tools/requirements.txt
CairoSVG==2.7.0
Pillow==9.5.0
